# apps/web/openapi/agent-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/ensure": {
      "post": {
        "operationId": "ensureUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnsureUserRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserRecordDto",
                      "nullable": true
                    },
                    "errorCode": {
                      "example": null,
                      "nullable": true,
                      "type": "string"
                    },
                    "errorMsg": {
                      "example": "",
                      "type": "string"
                    },
                    "errors": {
                      "additionalProperties": true,
                      "nullable": true,
                      "type": "object"
                    },
                    "success": {
                      "example": true,
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "errorMsg"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{userId}": {
      "get": {
        "operationId": "getUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserRecordDto",
                      "nullable": true
                    },
                    "errorCode": {
                      "example": null,
                      "nullable": true,
                      "type": "string"
                    },
                    "errorMsg": {
                      "example": "",
                      "type": "string"
                    },
                    "errors": {
                      "additionalProperties": true,
                      "nullable": true,
                      "type": "object"
                    },
                    "success": {
                      "example": true,
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "errorMsg"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Agent API",
    "description": "OpenAPI contract for the Agent workspace backend.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "data": {
            "type": "object",
            "additionalProperties": true,
            "example": null,
            "nullable": true
          },
          "errorMsg": {
            "type": "string",
            "example": "Request failed"
          },
          "errorCode": {
            "type": "string",
            "example": "BAD_REQUEST",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "path": {
            "type": "string",
            "example": "/api/conversations"
          },
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "timestamp": {
            "type": "string",
            "example": "2026-04-20T00:00:00.000Z"
          }
        },
        "required": [
          "success",
          "data",
          "errorMsg"
        ]
      },
      "UserRecordDto": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "createdAt",
          "displayName",
          "id",
          "updatedAt"
        ]
      },
      "EnsureUserRequestDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "default": "Local User",
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "displayName"
        ]
      },
...
    }
  }
}